get ranking by member

+++ Aftas/src/main/java/com/example/aftas/services/interfaces/RankingServiceInterface.java
@@ -14,7 +14,7 @@ public interface RankingServiceInterface {
 
     List<RankingResp> getRankingsByCompetitionCode(String competitionCode);
 
-    List<RankingResp> getRankingsByMemberNumber(int memberNumber);
+    List<RankingResp> getRankingsByMemberNumber(Long memberNumber);
 
     void calculateAndSetRankings(Competition competition);
